requirements.txt
awesome-slugify==1.6.5
bottle==0.12.8
CherryPy==3.6.0
Django==1.6
facebook-sdk==0.4.0
flake8==2.3.0
html5lib==0.99999
Jinja2==2.7.3
MarkupSafe==0.23
mccabe==0.3
pep8==1.6.2
Pillow==2.9.0
psycopg2==2.6
pyflakes==0.8.1
PyPDF2==1.24
regex==2015.6.4
reportlab==3.1.44
requests==2.5.1
six==1.9.0
Unidecode==0.4.17
xhtml2pdf==0.0.6
ujson
